requirements.txt
opencv-contrib-python
numpy
turbojpeg
nats-py